# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BiddingTrack": {
        "description": "FR-16.1 \u2014 what the eligibility engine turns on.\n\n`UNKNOWN` is a first-class answer, not a failure: NFR-LEGAL-1 requires we say\nso plainly rather than guess (eval C5 forbids confident-wrong here).",
        "enum": [
          "ncb",
          "icb",
          "donor",
          "private",
          "unknown"
        ],
        "title": "BiddingTrack",
        "type": "string"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807 problem+json. Every error from this API has this shape. Branch on `type` (a stable URL ending in the error code), never on `title` or on the prose in `detail`.",
        "properties": {
          "detail": {
            "description": "What went wrong this time. Prose; may change.",
            "type": "string"
          },
          "instance": {
            "description": "Request path the failure applies to.",
            "type": "string"
          },
          "status": {
            "description": "Repeats the HTTP status code.",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable summary of the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://adera.bid/errors/<code>` \u2014 the stable, matchable identity.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "TenderListOut": {
        "description": "Keyset-paginated page (05 \u00a712 \u2014 stable and O(1) at any depth, unlike OFFSET).\n\n`next_after` is an opaque cursor: pass it back as ?after= to get the next page;\nnull means last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TenderOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After"
          }
        },
        "required": [
          "items",
          "next_after"
        ],
        "title": "TenderListOut",
        "type": "object"
      },
      "TenderOut": {
        "description": "One tender, public fields only (facts + link-out \u2014 FR-9.1).",
        "properties": {
          "bid_bond_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Bond Currency"
          },
          "bid_bond_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Bond Minor"
          },
          "bidding_track": {
            "$ref": "#/components/schemas/BiddingTrack"
          },
          "buyer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer"
          },
          "closing_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing At"
          },
          "doc_price_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Price Currency"
          },
          "doc_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Price Minor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "opening_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening At"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "buyer",
          "summary",
          "region",
          "language",
          "url",
          "published_at",
          "closing_at",
          "opening_at",
          "bid_bond_minor",
          "bid_bond_currency",
          "doc_price_minor",
          "doc_price_currency",
          "bidding_track"
        ],
        "title": "TenderOut",
        "type": "object"
      },
      "TenderQAIn": {
        "description": "Question over a tender's parsed documents (FR-9.3).",
        "properties": {
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "TenderQAIn",
        "type": "object"
      },
      "TenderQAOut": {
        "description": "Answer with citations or refusal (FR-9.3, prompt B4).",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "tender_id": {
            "format": "uuid",
            "title": "Tender Id",
            "type": "string"
          }
        },
        "required": [
          "tender_id",
          "question",
          "answer",
          "citations",
          "confidence"
        ],
        "title": "TenderQAOut",
        "type": "object"
      },
...
      "ValidationProblem": {
        "description": "A ProblemDetail with per-field validation detail attached.",
        "properties": {
          "detail": {
            "description": "What went wrong this time. Prose; may change.",
            "type": "string"
          },
          "errors": {
            "description": "Pydantic's per-field errors: `loc`, `msg`, `type`.",
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "instance": {
            "description": "Request path the failure applies to.",
            "type": "string"
          },
          "status": {
            "description": "Repeats the HTTP status code.",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable summary of the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://adera.bid/errors/<code>` \u2014 the stable, matchable identity.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ValidationProblem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ADERA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tenders": {
      "get": {
        "operationId": "list_tenders_api_v1_tenders_get",
        "parameters": [
          {
            "description": "opaque cursor from next_after",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor from next_after",
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "List Tenders",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/search": {
      "get": {
        "operationId": "search_tenders_api_v1_tenders_search_get",
        "parameters": [
          {
            "description": "search query",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "search query",
              "title": "Q"
            }
          },
          {
            "description": "filter by region",
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "filter by region",
              "title": "Region"
            }
          },
          {
            "description": "filter by bidding track",
            "in": "query",
            "name": "bidding_track",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BiddingTrack"
                },
                {
                  "type": "null"
                }
              ],
              "description": "filter by bidding track",
              "title": "Bidding Track"
            }
          },
          {
            "description": "opaque cursor from next_after",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor from next_after",
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Search Tenders",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}": {
      "get": {
        "operationId": "get_tender_api_v1_tenders__tender_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`not_found` \u2014 no such resource \u2014 also returned instead of 403 across orgs, deliberately"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Get Tender",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}/qa": {
      "post": {
        "operationId": "tender_qa_api_v1_tenders__tender_id__qa_post",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenderQAIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderQAOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`not_found` \u2014 no such resource \u2014 also returned instead of 403 across orgs, deliberately"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Tender Qa",
        "tags": [
          "tenders"
        ]
      }
    },
...
  }
}